Add int32_t overload for ToMoney32FromGBP

+++ src/openrct2/world/Park.cpp
@@ -504,7 +504,7 @@ money64 Park::CalculateRideValue(const Ride* ride) const
     if (ride != nullptr && ride->value != RIDE_VALUE_UNDEFINED)
     {
         const auto& rtd = ride->GetRideTypeDescriptor();
-        result = ToMoney32FromGBP(static_cast<double long>(ride->value))
+        result = ToMoney32FromGBP(static_cast<int32_t>(ride->value))
             * (static_cast<money64>(ride_customers_in_last_5_minutes(ride)) + rtd.BonusValue * 4LL);
     }
     return result;
